feat: create common editor that can be used by api and graphql
