+++ app/client/src/components/editorComponents/GlobalSearch/SnippetsDescription.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 import SyntaxHighlighter from "react-syntax-highlighter/dist/cjs/prism-light";
 import sql from "react-syntax-highlighter/dist/cjs/languages/prism/sql";
 import { prism } from "react-syntax-highlighter/dist/cjs/styles/prism";
-import { TabbedViewContainer } from "pages/Editor/APIEditor/Form";
+import { TabbedViewContainer } from "pages/Editor/APIEditor/CommonEditorForm";
 import { TabComponent } from "components/ads/Tabs";
 import {
   EditorModes,
