+++ app/client/src/components/editorComponents/form/fields/EmbeddedDatasourcePathField.tsx
@@ -11,7 +11,6 @@ import CodeEditor, {
   EditorProps,
 } from "components/editorComponents/CodeEditor";
 import { CodeEditorBorder } from "components/editorComponents/CodeEditor/EditorConfig";
-import { API_EDITOR_FORM_NAME } from "constants/forms";
 import { AppState } from "reducers";
 import { connect } from "react-redux";
 import get from "lodash/get";
@@ -68,6 +67,7 @@ type ReduxStateProps = {
   applicationId?: string;
   dataTree: DataTree;
   actionName: string;
+  formName: string;
 };
 
 type ReduxDispatchProps = {
@@ -187,7 +187,6 @@ function CustomHint(props: { datasource: Datasource }) {
   );
 }
 
-const apiFormValueSelector = formValueSelector(API_EDITOR_FORM_NAME);
 class EmbeddedDatasourcePathComponent extends React.Component<
   Props,
   { highlightedElementWidth: number }
@@ -531,8 +530,9 @@ class EmbeddedDatasourcePathComponent extends React.Component<
 
 const mapStateToProps = (
   state: AppState,
-  ownProps: { pluginId: string; actionName: string },
+  ownProps: { pluginId: string; actionName: string; formName: string },
 ): ReduxStateProps => {
+  const apiFormValueSelector = formValueSelector(ownProps.formName);
   const datasourceFromAction = apiFormValueSelector(state, "datasource");
   let datasourceMerged = datasourceFromAction;
   // Todo: fix this properly later in #2164
@@ -558,12 +558,16 @@ const mapStateToProps = (
     applicationId: getCurrentApplicationId(state),
     dataTree: getDataTree(state),
     actionName: ownProps.actionName,
+    formName: ownProps.formName,
   };
 };
 
-const mapDispatchToProps = (dispatch: any): ReduxDispatchProps => ({
+const mapDispatchToProps = (
+  dispatch: any,
+  ownProps: any,
+): ReduxDispatchProps => ({
   updateDatasource: (datasource) =>
-    dispatch(change(API_EDITOR_FORM_NAME, "datasource", datasource)),
+    dispatch(change(ownProps.formName, "datasource", datasource)),
   setDatasourceEditorMode: (id: string, viewMode: boolean) =>
     dispatch(setDatsourceEditorMode({ id, viewMode })),
 });
@@ -580,6 +584,7 @@ function EmbeddedDatasourcePathField(
     theme: EditorTheme;
     actionName: string;
     codeEditorVisibleOverflow?: boolean;
+    formName: string;
   },
 ) {
   return (
