+++ app/client/src/pages/Editor/APIEditor/CommonEditorForm.tsx
@@ -1,11 +1,5 @@
 import React, { useState } from "react";
-import { connect, useDispatch, useSelector } from "react-redux";
-import {
-  change,
-  formValueSelector,
-  InjectedFormProps,
-  reduxForm,
-} from "redux-form";
+import { useDispatch, useSelector } from "react-redux";
 import {
   HTTP_METHOD_OPTIONS,
   API_EDITOR_TABS,
@@ -14,7 +8,6 @@ import styled from "styled-components";
 import FormLabel from "components/editorComponents/FormLabel";
 import FormRow from "components/editorComponents/FormRow";
 import { PaginationField, SuggestedWidget } from "api/ActionAPI";
-import { API_EDITOR_FORM_NAME } from "constants/forms";
 import Pagination from "./Pagination";
 import { Action, PaginationType } from "entities/Action";
 import {
@@ -22,13 +15,9 @@ import {
   toggleShowGlobalSearchModal,
 } from "actions/globalSearchActions";
 import KeyValueFieldArray from "components/editorComponents/form/fields/KeyValueFieldArray";
-import PostBodyData from "./PostBodyData";
-import ApiResponseView, {
-  EMPTY_RESPONSE,
-} from "components/editorComponents/ApiResponseView";
+import ApiResponseView from "components/editorComponents/ApiResponseView";
 import EmbeddedDatasourcePathField from "components/editorComponents/form/fields/EmbeddedDatasourcePathField";
 import { AppState } from "reducers";
-import { getApiName } from "selectors/formSelectors";
 import ActionNameEditor from "components/editorComponents/ActionNameEditor";
 import ActionSettings from "pages/Editor/ActionSettings";
 import RequestDropdownField from "components/editorComponents/form/fields/RequestDropdownField";
@@ -46,19 +35,13 @@ import {
   API_EDITOR_TAB_TITLES,
   createMessage,
   WIDGET_BIND_HELP,
-  API_PANE_NO_BODY,
 } from "@appsmith/constants/messages";
 import AnalyticsUtil from "utils/AnalyticsUtil";
 import CloseEditor from "components/editorComponents/CloseEditor";
 import { useParams } from "react-router";
-import get from "lodash/get";
 import DataSourceList from "./ApiRightPane";
 import { Datasource } from "entities/Datasource";
-import {
-  getAction,
-  getActionResponses,
-} from "../../../selectors/entitiesSelector";
-import { isEmpty, isEqual } from "lodash";
+import { isEqual } from "lodash";
 import { Colors } from "constants/Colors";
 import SearchSnippets from "components/ads/SnippetButton";
 import { ENTITY_TYPE } from "entities/DataTree/dataTreeFactory";
@@ -202,16 +185,6 @@ const TabSection = styled.div`
   overflow: auto;
 `;
 
-const NoBodyMessage = styled.div`
-  margin-top: 20px;
-  display: flex;
-  justify-content: center;
-
-  .${Classes.TEXT} {
-    color: ${(props) => props.theme.colors.apiPane.body.text};
-  }
-`;
-
 const CalloutContent = styled.div`
   display: flex;
   align-items: center;
@@ -232,7 +205,7 @@ const Wrapper = styled.div`
   height: calc(100% - 118px);
   position: relative;
 `;
-interface APIFormProps {
+export interface CommonFormProps {
   pluginId: string;
   onRunClick: (paginationField?: PaginationField) => void;
   onDeleteClick: () => void;
@@ -240,7 +213,6 @@ interface APIFormProps {
   isDeleting: boolean;
   paginationType: PaginationType;
   appName: string;
-  httpMethodFromForm: string;
   actionConfigurationHeaders?: any;
   actionConfigurationParams?: any;
   datasourceHeaders?: any;
@@ -261,7 +233,11 @@ interface APIFormProps {
   currentActionDatasourceId: string;
 }
 
-type Props = APIFormProps & InjectedFormProps<Action, APIFormProps>;
+type CommonFormPropsWithExtraParams = CommonFormProps & {
+  formName: string;
+  bodyUIComponent: JSX.Element;
+  handleSubmit: any;
+};
 
 export const NameWrapper = styled.div`
   display: flex;
@@ -526,7 +502,7 @@ function ImportedDatas(props: { data: any; attributeName: string }) {
   );
 }
 
-function ApiEditorForm(props: Props) {
+function CommonEditorForm(props: CommonFormPropsWithExtraParams) {
   const [selectedIndex, setSelectedIndex] = useState(0);
   const [
     apiBindHelpSectionVisible,
@@ -538,10 +514,9 @@ function ApiEditorForm(props: Props) {
     actionConfigurationParams,
     actionName,
     currentActionDatasourceId,
-    handleSubmit,
+    formName,
     headersCount,
     hintMessages,
-    httpMethodFromForm,
     isRunning,
     onRunClick,
     paramsCount,
@@ -550,7 +525,6 @@ function ApiEditorForm(props: Props) {
     updateDatasource,
   } = props;
   const dispatch = useDispatch();
-  const allowPostBody = httpMethodFromForm;
 
   const params = useParams<{ apiId?: string; queryId?: string }>();
 
@@ -576,7 +550,7 @@ function ApiEditorForm(props: Props) {
   return (
     <>
       <CloseEditor />
-      <Form onSubmit={handleSubmit}>
+      <Form onSubmit={props.handleSubmit}>
         <MainConfiguration>
           <FormRow className="form-row-header">
             <NameWrapper className="t--nameOfApi">
@@ -624,6 +598,7 @@ function ApiEditorForm(props: Props) {
               <EmbeddedDatasourcePathField
                 actionName={actionName}
                 codeEditorVisibleOverflow
+                formName={formName}
                 name="actionConfiguration.path"
                 placeholder="https://mock-api.appsmith.com/users"
                 pluginId={pluginId}
@@ -703,18 +678,7 @@ function ApiEditorForm(props: Props) {
                   {
                     key: API_EDITOR_TABS.BODY,
                     title: createMessage(API_EDITOR_TAB_TITLES.BODY),
-                    panelComponent: allowPostBody ? (
-                      <PostBodyData
-                        dataTreePath={`${actionName}.config`}
-                        theme={theme}
-                      />
-                    ) : (
-                      <NoBodyMessage>
-                        <Text type={TextType.P2}>
-                          {createMessage(API_PANE_NO_BODY)}
-                        </Text>
-                      </NoBodyMessage>
-                    ),
+                    panelComponent: props.bodyUIComponent,
                   },
                   {
                     key: API_EDITOR_TABS.PAGINATION,
@@ -730,7 +694,7 @@ function ApiEditorForm(props: Props) {
                   {
                     key: API_EDITOR_TABS.AUTHENTICATION,
                     title: createMessage(API_EDITOR_TAB_TITLES.AUTHENTICATION),
-                    panelComponent: <ApiAuthentication />,
+                    panelComponent: <ApiAuthentication formName={formName} />,
                   },
                   {
                     key: API_EDITOR_TABS.SETTINGS,
@@ -739,7 +703,7 @@ function ApiEditorForm(props: Props) {
                       <SettingsWrapper>
                         <ActionSettings
                           actionSettingsConfig={settingsConfig}
-                          formName={API_EDITOR_FORM_NAME}
+                          formName={formName}
                           theme={theme}
                         />
                       </SettingsWrapper>
@@ -770,110 +734,4 @@ function ApiEditorForm(props: Props) {
   );
 }
 
-const selector = formValueSelector(API_EDITOR_FORM_NAME);
-
-type ReduxDispatchProps = {
-  updateDatasource: (datasource: Datasource) => void;
-};
-
-const mapDispatchToProps = (dispatch: any): ReduxDispatchProps => ({
-  updateDatasource: (datasource) => {
-    dispatch(change(API_EDITOR_FORM_NAME, "datasource", datasource));
-  },
-});
-
-export default connect((state: AppState, props: { pluginId: string }) => {
-  const httpMethodFromForm = selector(state, "actionConfiguration.httpMethod");
-  const actionConfigurationHeaders =
-    selector(state, "actionConfiguration.headers") || [];
-  const actionConfigurationParams =
-    selector(state, "actionConfiguration.queryParameters") || [];
-  let datasourceFromAction = selector(state, "datasource");
-  if (datasourceFromAction && datasourceFromAction.hasOwnProperty("id")) {
-    datasourceFromAction = state.entities.datasources.list.find(
-      (d) => d.id === datasourceFromAction.id,
-    );
-  }
-
-  // get messages from action itself
-  const actionId = selector(state, "id");
-  const action = getAction(state, actionId);
-  const hintMessages = action?.messages;
-
-  const datasourceHeaders =
-    get(datasourceFromAction, "datasourceConfiguration.headers") || [];
-  const datasourceParams =
-    get(datasourceFromAction, "datasourceConfiguration.queryParameters") || [];
-
-  const apiId = selector(state, "id");
-  const currentActionDatasourceId = selector(state, "datasource.id");
-
-  const actionName = getApiName(state, apiId) || "";
-  const headers = selector(state, "actionConfiguration.headers");
-  let headersCount = 0;
-
-  if (Array.isArray(headers)) {
-    const validHeaders = headers.filter(
-      (value) => value.key && value.key !== "",
-    );
-    headersCount += validHeaders.length;
-  }
-
-  if (Array.isArray(datasourceHeaders)) {
-    const validHeaders = datasourceHeaders.filter(
-      (value: any) => value.key && value.key !== "",
-    );
-    headersCount += validHeaders.length;
-  }
-
-  const params = selector(state, "actionConfiguration.queryParameters");
-  let paramsCount = 0;
-
-  if (Array.isArray(params)) {
-    const validParams = params.filter((value) => value.key && value.key !== "");
-    paramsCount = validParams.length;
-  }
-
-  if (Array.isArray(datasourceParams)) {
-    const validParams = datasourceParams.filter(
-      (value: any) => value.key && value.key !== "",
-    );
-    paramsCount += validParams.length;
-  }
-
-  const responses = getActionResponses(state);
-  let hasResponse = false;
-  let suggestedWidgets;
-  if (apiId && apiId in responses) {
-    const response = responses[apiId] || EMPTY_RESPONSE;
-    hasResponse =
-      !isEmpty(response.statusCode) && response.statusCode[0] === "2";
-    suggestedWidgets = response.suggestedWidgets;
-  }
-
-  return {
-    actionName,
-    apiId,
-    httpMethodFromForm,
-    actionConfigurationHeaders,
-    actionConfigurationParams,
-    currentActionDatasourceId,
-    datasourceHeaders,
-    datasourceParams,
-    headersCount,
-    paramsCount,
-    hintMessages,
-    datasources: state.entities.datasources.list.filter(
-      (d) => d.pluginId === props.pluginId,
-    ),
-    currentPageId: state.entities.pageList.currentPageId,
-    applicationId: state.entities.pageList.applicationId,
-    suggestedWidgets,
-    hasResponse,
-  };
-}, mapDispatchToProps)(
-  reduxForm<Action, APIFormProps>({
-    form: API_EDITOR_FORM_NAME,
-    enableReinitialize: true,
-  })(ApiEditorForm),
-);
+export default CommonEditorForm;
